+++ api/api.test.js
@@ -15,9 +15,9 @@ tap.test('express api', async t => {
     api = require('./api');
   });
 
-  t.test('GET /heartbeat', async t => {
+  t.test('GET /heartbeat', async test => {
     response = await request(api).get('/heartbeat');
-    t.equal(response.status, 204, 'HTTP status set to 204');
+    test.equal(response.status, 204, 'HTTP status set to 204');
   });
 
   // t.test('GET /api-docs', async t => {
@@ -25,51 +25,54 @@ tap.test('express api', async t => {
   //   t.equals(response.status, 301, 'successful');
   // });
 
-  t.test('headers', async t => {
+  t.test('headers', async test => {
     response = await request(api).get('/');
-    t.notOk(response.header['x-powered-by'], 'X-Powered-By header is unset');
-    t.equal(
+    test.notOk(response.header['x-powered-by'], 'X-Powered-By header is unset');
+    test.equal(
       response.header['cache-control'],
       'private, no-cache',
       'Cache-Control header is set'
     );
-    t.ok(
+    test.ok(
       response.header['strict-transport-security'],
       'Strict-Transport-Security header is set'
     );
-    t.equal(
+    test.equal(
       response.header['x-content-type-options'],
       'nosniff',
       'X-Content-Type-Options header is set'
     );
-    t.equal(
+    test.equal(
       response.header['x-frame-options'],
       'sameorigin',
       'X-Frame-Options is set'
     );
-    t.equal(
+    test.equal(
       response.header['x-xss-protection'],
       '1; mode=block',
       'X-XSS-Protection is set'
     );
   });
 
-  t.test('error handling', async t => {
+  t.test('error handling', async test => {
     // disconnect database from api via environment
-    t.beforeEach(async () => {
+    test.beforeEach(async () => {
       testDbHost = process.env.TEST_DB_HOST;
       process.env.TEST_DB_HOST = 'undefined';
       api = require('./api');
     });
 
     // reattach database
-    t.afterEach(async () => {
+    test.afterEach(async () => {
       process.env.TEST_DB_HOST = testDbHost;
     });
 
-    t.test('GET /states without a database connection returns 500', async t => {
-      const response = await request(api).get('/states');
-      t.equal(response.status, 500, 'HTTP status set to 500');
-    });
+    test.test(
+      'GET /states without a database connection returns 500',
+      async t => {
+        const response = await request(api).get('/states');
+        t.equal(response.status, 500, 'HTTP status set to 500');
+      }
+    );
   });
 });